requirements.txt
gradio==5.23.1
Pillow==11.1.0